requirements.txt
streamlit==1.38.0
openai==1.44.0
python-dotenv==1.0.1
requests==2.32.3
faiss-cpu==1.8.0
toml==0.10.2
httpx==0.27.2